feat: implement clear cache refact: move all resources to one CustomerResource class
