+++ crm-customer-service-adapters/spring-data-repository/build.gradle
@@ -21,6 +21,8 @@ dependencies {
 
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
 
+    implementation 'javax.inject:javax.inject:1'
+
     implementation 'org.flywaydb:flyway-core'
     runtimeOnly 'org.postgresql:postgresql'
