+++ crm-customer-service-core/build.gradle
@@ -18,6 +18,8 @@ repositories {
 }
 
 dependencies {
+    implementation 'javax.inject:javax.inject:1'
+    implementation 'org.projectlombok:lombok:1.18.20'
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.2'
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.2'
     testCompileOnly('org.junit.jupiter:junit-jupiter-params:5.8.2')
