+++ crm-customer-service-core/src/main/java/br/com/crmcustomer/core/port/CustomerRepository.java
@@ -16,4 +16,5 @@ public interface CustomerRepository {
 
     void deleteCustomer(String documentContent);
 
+    void clearCache();
 }
