+++ crm-customer-service-core/src/main/java/br/com/crmcustomer/core/usecase/clearcache/ClearCacheUseCase.java
@@ -1,0 +1,5 @@
+package br.com.crmcustomer.core.usecase.clearcache;
+
+public interface ClearCacheUseCase {
+    void clearCache();
+}
