+++ crm-customer-service-core/src/test/java/br/com/crmcustomer/core/repository/CustomerTestRepository.java
@@ -30,4 +30,9 @@ public class CustomerTestRepository implements CustomerRepository {
     public void deleteCustomer(String documentContent) {
 
     }
+
+    @Override
+    public void clearCache() {
+
+    }
 }
